Renamed psr-4 directory in lib
