+++ composer.json
@@ -10,7 +10,7 @@
     },
     "autoload": {
         "psr-4": {
-            "PhpLogger\\" : "Logger/"
+            "PhpLogger\\" : "lib/"
         }
     },
     "license": "MIT",
