+++ lib/Log/LogInterface.php
@@ -8,7 +8,7 @@
  * @package Log
  */
 
-namespace PhpLogger\Log;
+namespace lib\Log;
 
 interface LogInterface extends \JsonSerializable
 {
